examples/list_tools/requirements.txt
fastmcp_transport_ble==0.0.1a1
